crawl-engine/requirements.txt
aiohttp>=3.9
aiosqlite>=0.19
imagehash>=4.3
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.65
python-dotenv>=1.0
# Optional: for R2 storage
# boto3>=1.28
# Optional: for AI classification
# open-clip-torch>=2.20
# torch>=2.0
